move e2e tests to separate crate to prevent circular dependency between test-utils and telcoin-network binary
