--- bin/telcoin-network/tests/it/main.rs
@@ -1,9 +1,0 @@
-//! CLI integration test
-
-#[cfg(feature = "faucet")]
-mod faucet;
-mod genesis_tests;
-mod restarts;
-mod util;
-
-async fn main() {}
