+++ tests/e2e-tests/Cargo.toml
@@ -1,0 +1,70 @@
+[package]
+name = "e2e-tests"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+authors.workspace = true
+description.workspace = true
+exclude.workspace = true
+
+[dependencies]
+# utils
+telcoin-network = { path = "../../bin/telcoin-network" }
+tn-test-utils = { workspace = true }
+tn-node = { workspace = true }
+tempfile = { workspace = true }
+clap = { workspace = true, features = ["derive", "env"] }
+reth-chainspec = { workspace = true }
+tracing = { workspace = true }
+eyre = { workspace = true }
+
+# faucet
+tokio = { workspace = true, features = ["time"], optional = true }
+jsonrpsee = { workspace = true, optional = true }
+tn-faucet = { workspace = true, optional = true }
+gcloud-sdk = { version = "=0.24.6", default-features = false, features = [
+    "google-cloud-kms-v1",
+    "tls-webpki-roots",
+], optional = true }
+k256 = { version = "0.13.3", optional = true }
+tn-types = { workspace = true, optional = true }
+futures = { workspace = true, optional = true }
+
+[dev-dependencies]
+alloy = { workspace = true }
+serde_json = { workspace = true }
+rand = { workspace = true }
+secp256k1 = { workspace = true, features = [
+    "global-context",
+    "rand-std",
+    "recovery",
+] }
+tn-types = { workspace = true }
+tn-config = { workspace = true }
+jsonrpsee = { workspace = true }
+ethereum-tx-sign = "6.1.3"
+const-hex = "1.12.0"
+# Alternate sync json rpc lib- just for integration testing.
+jsonrpc = "0.18.0"
+nix = { version = "0.29", features = ["signal"] }
+tokio = { workspace = true, features = [
+    "sync",
+    "macros",
+    "time",
+    "rt-multi-thread",
+] }
+
+[features]
+default = []
+faucet = [
+    "tn-faucet",
+    "tokio",
+    "jsonrpsee",
+    "dep:gcloud-sdk",
+    "dep:k256",
+    "dep:futures",
+    "tn-types",
+]
